dynamically loaded and instanced game states.  broke combat.
