+++ tuxemon/core/components/event/actions/player.py
@@ -59,10 +59,9 @@ class Player(object):
         monster = game.imports["monster"]
         Map = game.imports["map"].Map
 
-
         # Get the player object from the game.
         player = game.player1
-        world = game.state_dict["WORLD"]
+        world = game.state
 
         # Get the teleport parameters for the position x,y and the map to load.
         parameters = action[1].split(",")
@@ -120,12 +119,12 @@ class Player(object):
                                     layer_pos += 1
                             y_pos += 1
                         x_pos += 1
-        
+
         # Update the server/clients of our new map and populate any other players.
         if game.isclient or game.ishost:
             game.add_clients_to_map(game.client.client.registry)
             game.client.update_player(player.facing)
-            
+
         # Stop the player's movement so they don't continue their move after they teleported.
         player.moving = False
 
@@ -295,8 +294,8 @@ class Player(object):
 
         # If we're doing a transition, only change the player's facing when we've reached the apex
         # of the transition.
-        if game.state_dict["WORLD"].start_transition:
-            game.state_dict["WORLD"].delayed_facing = parameters
+        if game.state.start_transition:
+            game.state.delayed_facing = parameters
         else:
             game.player1.facing = parameters
