+++ tuxemon/core/states/multiplayer/__init__.py
@@ -34,13 +34,15 @@ import os
 import sys
 import pprint
 
-from .. import tools, prepare
+from core import prepare
+from core import state
 
 # Create a logger for optional handling of debug messages.
 logger = logging.getLogger(__name__)
 logger.debug("states.start successfully imported")
 
-class Headless(tools._State):
+
+class Headless(state.State):
     """The module responsible for the headless server.
 
     :param game: The scene manager object that contains all the game's variables.
@@ -49,21 +51,21 @@ class Headless(tools._State):
     """
 
     def __init__(self, game):
-        
+
         # Initiate our common state properties.
-        tools._State.__init__(self)
-        
+        state.State.__init__(self)
+
         self.game = game
-        
+
         # Set the native tile size so we know how much to scale
         self.tile_size = prepare.TILE_SIZE
- 
+
         # Set the status icon size so we know how much to scale
         self.icon_size = prepare.ICON_SIZE
- 
+
         # Get the screen's resolution
         self.resolution = prepare.SCREEN_SIZE
- 
+
         # Native resolution is similar to the old gameboy resolution. This is
         # used for scaling.
         self.native_resolution = prepare.NATIVE_RESOLUTION
@@ -72,10 +74,9 @@ class Headless(tools._State):
         # Set the tiles and mapsize variables
         self.tiles = []
         self.map_size = []
-        
+
         self.global_x = 0
         self.global_y = 0
-        
+
         self.npcs = []
         self.npcs_off_map = []
-
