+++ tuxemon/core/states/pc/__init__.py
@@ -31,22 +31,17 @@
 """
 import logging
 import pygame
-import os
-import sys
-import pprint
-import random
 
-from .. import tools, prepare
-from ..components import pyganim
-from ..components import db
-from ..components import fusion
-from ..components.menu import pc_menu
+from core import prepare
+from core import state
+from core.components.menu import pc_menu
 
 # Create a logger for optional handling of debug messages.
 logger = logging.getLogger(__name__)
 logger.debug("states.start successfully imported")
 
-class PC(tools._State):
+
+class PC(state.State):
     """The module responsible in game settings.
 
     :param game: The scene manager object that contains all the game's variables.
@@ -56,15 +51,15 @@ class PC(tools._State):
 
     def __init__(self, game):
         # Initiate our common state properties.
-        tools._State.__init__(self)
-    
-        from ..components import menu
-        
+        state.State.__init__(self)
+
+        from core.components import menu
+
         # Provide an instance of the scene manager to this scene.
         self.game = game            # The scene manger object
         self.previous_menu = None
         self.menu_blocking = True
-        
+
          # Provide access to the screen surface
         self.screen = game.screen
         self.screen_rect = prepare.SCREEN_RECT
@@ -82,15 +77,15 @@ class PC(tools._State):
         # used for scaling.
         self.native_resolution = prepare.NATIVE_RESOLUTION
         self.scale = prepare.SCALE
-        
+
         # Main PC menu.
         self.pc_menu = pc_menu.PCMenu(self.screen,
                                       self.resolution,
                                       self.game)
-        
+
         self.pc_menu.interactable = True
         self.pc_menu.size_ratio = [0.8, 0.3]
-        
+
         # Main multiplayer menu.
         self.multiplayer_menu = pc_menu.Multiplayer_Menu(self.screen,
                                                          self.resolution,
@@ -98,7 +93,7 @@ class PC(tools._State):
         self.multiplayer_menu.visible = False
         self.multiplayer_menu.interactable = False
         self.multiplayer_menu.size_ratio = [0.7, 0.25]
-        
+
         # Join a multiplayer game menu.
         self.multiplayer_join_menu = pc_menu.Multiplayer_Join_Menu(self.screen,
                                                                    self.resolution,
@@ -106,7 +101,7 @@ class PC(tools._State):
         self.multiplayer_join_menu.visible = False
         self.multiplayer_join_menu.interactable = False
         self.multiplayer_join_menu.size_ratio = [0.6, 0.2]
-        
+
         # Successfully joined a multiplayer game menu.
         self.multiplayer_join_success_menu = pc_menu.Multiplayer_Join_Success_Menu(self.screen,
                                                                                    self.resolution,
@@ -114,7 +109,7 @@ class PC(tools._State):
         self.multiplayer_join_success_menu.visible = False
         self.multiplayer_join_success_menu.interactable = False
         self.multiplayer_join_success_menu.size_ratio = [0.6, 0.2]
-        
+
         # Successfully host a game menu.
         self.multiplayer_host_menu = pc_menu.Multiplayer_Host_Menu(self.screen,
                                                                    self.resolution,
@@ -122,15 +117,14 @@ class PC(tools._State):
         self.multiplayer_host_menu.visible = False
         self.multiplayer_host_menu.interactable = False
         self.multiplayer_host_menu.size_ratio = [0.6, 0.2]
-        
-        self.menus = [self.pc_menu, 
-                      self.multiplayer_menu, 
-                      self.multiplayer_join_menu, 
-                      self.multiplayer_join_success_menu, 
+
+        self.menus = [self.pc_menu,
+                      self.multiplayer_menu,
+                      self.multiplayer_join_menu,
+                      self.multiplayer_join_success_menu,
                       self.multiplayer_host_menu
                       ]
-        
-        
+
         for menu in self.menus:
             menu.scale = self.scale    # Set the scale of the menu.
             menu.set_font(size=menu.font_size * self.scale,
@@ -150,14 +144,14 @@ class PC(tools._State):
                     border,
                     (border.get_width() * self.scale,
                      border.get_height() * self.scale))
-        
+
             # Set the menu size.
             menu.size_x= int(self.resolution[0] * menu.size_ratio[0])
             menu.size_y= int(self.resolution[1] * menu.size_ratio[1])
             menu.pos_x = (self.resolution[0] / 2) - (menu.size_x/2)
             menu.pos_y = (self.resolution[1] / 2) - (menu.size_y/2)
-    
-        
+
+
     def startup(self, current_time, persistant):
         """Perform startup tasks when we switch to this scene.
 
@@ -179,7 +173,7 @@ class PC(tools._State):
         {}
 
         """
-        
+
         self.persist = persistant
         self.start_time = current_time
 
@@ -242,20 +236,20 @@ class PC(tools._State):
         if self.multiplayer_join_success_menu.interactable:
             if event.type == pygame.KEYDOWN and event.key == pygame.K_ESCAPE:
                 self.game.get_menu_event(self.multiplayer_join_success_menu, event)
-        
+
         elif self.multiplayer_host_menu.interactable:
             if event.type == pygame.KEYDOWN and event.key == pygame.K_ESCAPE:
                 self.game.get_menu_event(self.multiplayer_host_menu, event)
-            
+
         elif self.multiplayer_join_menu.interactable:
             self.game.get_menu_event(self.multiplayer_join_menu, event)
-        
+
         elif self.multiplayer_menu.interactable:
             self.game.get_menu_event(self.multiplayer_menu, event)
-        
+
         elif self.pc_menu.interactable:
             self.game.get_menu_event(self.pc_menu, event)
-            
+
 
     def draw(self):
         """Draws the start screen to the screen.
@@ -272,33 +266,33 @@ class PC(tools._State):
         self.pc_menu.draw()
         self.pc_menu.draw_textItem(
                 ["MULTIPLAYER", "LOG OFF"])
-        
-        
+
+
         if self.multiplayer_menu.visible:
             self.multiplayer_menu.draw()
             self.multiplayer_menu.draw_textItem(
                 ["JOIN", "HOST"])
 
-        
+
         if self.multiplayer_join_menu.visible:
             self.multiplayer_join_menu.draw()
             self.multiplayer_join_menu.draw_textItem(self.game.client.server_list)
-            
+
             # If no options are selected because there were no items when the menu was populated,
-            # and there are items in the list to select, set the selected item to the top of the list. 
+            # and there are items in the list to select, set the selected item to the top of the list.
             if self.multiplayer_join_menu.selected_menu_item <= 0 and \
             len(self.multiplayer_join_menu.menu_items) > 0:
                 self.multiplayer_join_menu.selected_menu_item = 0
-        
+
         if self.multiplayer_join_success_menu.visible:
             self.multiplayer_join_success_menu.draw()
-            self.multiplayer_join_success_menu.draw_textItem(self.multiplayer_join_success_menu.text)   
-        
+            self.multiplayer_join_success_menu.draw_textItem(self.multiplayer_join_success_menu.text)
+
         if self.multiplayer_host_menu.visible:
             self.multiplayer_host_menu.draw()
-            self.multiplayer_host_menu.draw_textItem(self.multiplayer_host_menu.text)  
-            
-    
-    
+            self.multiplayer_host_menu.draw_textItem(self.multiplayer_host_menu.text)
+
+
+
