+++ tuxemon/tuxemon.cfg
@@ -13,7 +13,7 @@ controller_transparency = 45
 starting_map = bedroom_test.tmx
 starting_position_x = 3
 starting_position_y = 3
-cli_enabled = 0
+cli_enabled = 1
 net_controller_enabled = 1
 
 [player]
